requirements.txt
Django>=4.2,<5.1
wagtail>=6.2,<6.3
whitenoise == 6.6.0
django-environ == 0.11.2
psycopg2-binary == 2.9.9
gunicorn == 22.0.0